hostinger_deploy/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
SQLAlchemy==2.0.23
email-validator==2.1.0
Pillow==10.1.0
numpy==1.26.2
selenium==4.15.2
psycopg2-binary==2.9.9
gunicorn==21.2.0
